Move enum types to `common.types` module
They were buried deep within the `workflows.relax.generator` module but
most of them will be useful outside of the context of just the relax
workflow as well. It makes more sense to place them in the common
module instead.
