+++ aiida_common_workflows/common/__init__.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+# pylint: disable=redefined-builtin,undefined-variable
+"""Module for resources common to the entire `aiida-common-workflows` package."""
+from .types import *
+
+all = (types.__all__,)
