--- aiida_common_workflows/workflows/relax/README.md
@@ -1,3 +1,0 @@
-# Relaxation WorkChain common interface
-
-For info, look at the page "Relaxation of a structure" in the wiki of this repository.
