+++ aiida_common_workflows/workflows/relax/bigdft/generator.py
@@ -7,7 +7,8 @@ from aiida import orm
 from aiida import plugins
 from aiida.engine import calcfunction
 
-from ..generator import RelaxInputGenerator, RelaxType, SpinType, ElectronicType
+from aiida_common_workflows.common import ElectronicType, RelaxType, SpinType
+from ..generator import RelaxInputGenerator
 
 __all__ = ('BigDftRelaxInputGenerator',)
