+++ aiida_common_workflows/workflows/relax/cp2k/generator.py
@@ -10,7 +10,8 @@ from aiida import engine
 from aiida import orm
 from aiida import plugins
 
-from ..generator import RelaxInputGenerator, RelaxType, SpinType, ElectronicType
+from aiida_common_workflows.common import ElectronicType, RelaxType, SpinType
+from ..generator import RelaxInputGenerator
 
 __all__ = ('Cp2kRelaxInputGenerator',)
