+++ aiida_common_workflows/workflows/relax/quantum_espresso/generator.py
@@ -6,7 +6,8 @@ from aiida import engine
 from aiida import orm
 from aiida import plugins
 
-from ..generator import RelaxInputGenerator, RelaxType, SpinType, ElectronicType
+from aiida_common_workflows.common import ElectronicType, RelaxType, SpinType
+from ..generator import RelaxInputGenerator
 
 __all__ = ('QuantumEspressoRelaxInputGenerator',)
