+++ aiida_common_workflows/workflows/relax/siesta/generator.py
@@ -9,7 +9,9 @@ from aiida import engine
 from aiida import orm
 from aiida import plugins
 from aiida.common import exceptions
-from ..generator import RelaxInputGenerator, RelaxType, SpinType, ElectronicType
+
+from aiida_common_workflows.common import ElectronicType, RelaxType, SpinType
+from ..generator import RelaxInputGenerator
 
 __all__ = ('SiestaRelaxInputGenerator',)
