+++ tests/workflows/relax/test_generator.py
@@ -4,7 +4,8 @@
 import pytest
 
 from aiida_common_workflows.protocol import ProtocolRegistry
-from aiida_common_workflows.workflows.relax import RelaxInputGenerator, RelaxType, SpinType, ElectronicType
+from aiida_common_workflows.common import RelaxType, SpinType, ElectronicType
+from aiida_common_workflows.workflows.relax import RelaxInputGenerator
 from aiida_common_workflows.workflows.relax.workchain import CommonRelaxWorkChain
